itest: move recovery related tests into one file
